+++ R/pkg.R
@@ -1,5 +1,5 @@
 #' @importFrom Rcpp evalCpp
-#' @importFrom data.table tstrsplit ":=" data.table as.data.table setDF setDT setnames rbindlist setorder
+#' @importFrom data.table tstrsplit ":=" data.table as.data.table setDF setDT setnames rbindlist setorder is.data.table
 #' @importFrom utils download.file head
 #' @importFrom Matrix sparseMatrix summary crossprod tcrossprod diag 
 #' @useDynLib udpipe