+++ NEWS.md
@@ -20,6 +20,7 @@
 - Added predict.LDA_VEM and predict.LDA_Gibbs
 - Renamed dataset annotation_params to udpipe_annotation_params
 - Added example datasets called brussels_listings, brussels_reviews, brussels_reviews_anno
+- Use path.expand on conll-u files which are used for training
 
 # CHANGES IN udpipe VERSION 0.1.2